+++ packages/database/src/sql/index.ts
@@ -1,81 +1,11 @@
 import { DatabaseType } from 'typeorm'
-import { BaseSqlInterface } from './base'
+import { config } from '../config'
+import { Mysql } from './mysql'
 
-const coreCmdMap: Partial<Record<DatabaseType, string>> = {
-  mysql: 'mysql',
-  postgres: 'psql',
+const getSql = (type: DatabaseType) => {
+  if (type === 'mysql') return new Mysql()
+  if (type === 'postgres') return new Mysql()
+  return new Mysql()
 }
 
-const dumpCmdMap: Partial<Record<DatabaseType, string>> = {
-  mysql: 'mysqldump',
-  postgres: 'pg_dump',
-}
-
-export class BaseSql implements BaseSqlInterface {
-  type: DatabaseType
-  coreCmd: string
-  dumpCmd: string
-
-  /** instantiate a new Sql class */
-  constructor(type: DatabaseType) {
-    this.type = type
-    this.coreCmd = coreCmdMap[type]
-    this.dumpCmd = dumpCmdMap[type]
-  }
-
-  /**
-   * removes a single table from a mysql database
-   */
-  async dropTable(database: string, table: string) {
-    console.log('params', database, table)
-    console.log('Not implemented yet')
-  }
-
-  /**
-   * removes a list of tables from a mysql database
-   */
-  async dropTables(database: string, tables: string[]) {
-    console.log('params', database, tables)
-    console.log('Not implemented yet')
-  }
-
-  /**
-   * drops the database
-   */
-  async dropDatabase(database: string) {
-    console.log('params', database)
-    console.log('Not implemented yet')
-  }
-
-  /**
-   * a very aggressive function that drops the database
-   * and then recreates it for a completely fresh start
-   * so ensure .env.test has the corrent database name.
-   * @param {string} database
-   */
-  async clearDatabase(database: string) {
-    console.log('params', database)
-    console.log('Not implemented yet')
-  }
-
-  /**
-   * restores SQL database from a file
-   * @param {string} database
-   * @param {string} filename
-   */
-  async restoreFromFile(database: string, filename: string) {
-    console.log('params', database, filename)
-    console.log('Not implemented yet')
-  }
-
-  /** interactive prompt to guide the user to restore an SQL database */
-  restorePrompted(database: string) {
-    console.log('params', database)
-    console.log('Not implemented yet')
-  }
-
-  async dump(database: string) {
-    console.log('params', database)
-    console.log('Not implemented yet')
-  }
-}
+export const sql = getSql(config.type)
